% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_ariel}
\alias{theme_ariel}
\title{Theme Ariel}
\usage{
theme_ariel(base_size = 12, base_family = "ibm")
}
\arguments{
\item{base_size}{base font size, in pts}

\item{base_family}{base font family, using sysfonts alias}
}
\value{
A ggplot2 object
}
\description{
This function initializes a new theme with custom settings.
}
